+++ components/shared/AppSidebar/InsightsPanel.tsx
@@ -10,6 +10,7 @@ interface InsightsPanelProps {
   insights: DbUserList[] | DbUserInsight[];
   type: "repo" | "list";
   isLoading: boolean;
+  workspaceId?: string | null;
 }
 
 const Loading = () => {
@@ -26,7 +27,7 @@ const Loading = () => {
   );
 };
 
-export const InsightsPanel = ({ title, username, insights, type, isLoading }: InsightsPanelProps) => {
+export const InsightsPanel = ({ title, username, insights, type, isLoading, workspaceId }: InsightsPanelProps) => {
   const [open, setOpen] = useState(true);
   const getIcon = (type: "repo" | "list") => {
     switch (type) {
@@ -38,13 +39,28 @@ export const InsightsPanel = ({ title, username, insights, type, isLoading }: In
   };
   return (
     <div>
-      <SidebarMenuItem title={title} url={type === "list" ? "/hub/lists" : "/hub/insights"} icon={getIcon(type)} />
+      <SidebarMenuItem
+        title={title}
+        url={
+          type === "list"
+            ? workspaceId
+              ? `/workspaces/${workspaceId}/contributor-insights`
+              : "/hub/lists"
+            : "/hub/insights"
+        }
+        icon={getIcon(type)}
+      />
 
       <div className="overflow-hidden">
         {isLoading ? null : (
           <ul className="list-none w-full px-4 mt-1 [&_li]:border-l-2 text-slate-700 tracking-tight">
             {insights.slice(0, 3).map((insight) => {
-              const url = type === "list" ? `/lists/${insight.id}` : `/pages/${username}/${insight.id}/dashboard`;
+              const url =
+                type === "list"
+                  ? workspaceId
+                    ? `/workspaces/${workspaceId}/contributor-insights/${insight.id}/overview`
+                    : `/lists/${insight.id}`
+                  : `/pages/${username}/${insight.id}/dashboard`;
               return (
                 <li
                   className="py-1 px-3 hover:bg-slate-100 rounded-tr-md rounded-br-md transition-colors text-sm"
@@ -62,7 +78,13 @@ export const InsightsPanel = ({ title, username, insights, type, isLoading }: In
               <Link
                 className="text-xs text-slate-500 pl-3 pt-2 border-l-2 hover:text-orange-600"
                 title=""
-                href={type === "list" ? "/hub/lists" : "/hub/insights"}
+                href={
+                  type === "list"
+                    ? workspaceId
+                      ? `/workspaces/${workspaceId}/contributor-insights`
+                      : "/hub/lists"
+                    : "/hub/insights"
+                }
               >
                 Show all
               </Link>